content_gen/requirements.txt
openai>=1.0.0
elevenlabs>=0.2.24
pydub>=0.25.1
requests>=2.31.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
flask>=2.0.0
flask-cors>=3.0.0